implemented member system

+++ packages/backend/src/rest/comments/comment.controller.ts
@@ -1,7 +1,10 @@
-import { Body, Controller, Delete, Get, Param, Post, Put, Query, UseGuards } from '@nestjs/common'
+import { Body, Controller, Delete, ForbiddenException, forwardRef, Get, Inject, NotFoundException, Param, Post, Put, Query, UseGuards } from '@nestjs/common'
+import { REQUEST } from '@nestjs/core'
 import { AuthGuard } from '@nestjs/passport'
 import { ApiBasicAuth, ApiBody, ApiParam, ApiQuery, ApiResponse } from '@nestjs/swagger'
-import { Comment, CommentAddData, CommentUpdateData, CommentREST } from 'productboard-common'
+import { Comment, CommentAddData, CommentUpdateData, CommentREST, User } from 'productboard-common'
+import { IssueService } from '../issues/issue.service'
+import { MemberService } from '../members/member.service'
 import { CommentService } from './comment.service'
 
 @Controller('rest/comments')
@@ -9,7 +12,12 @@ import { CommentService } from './comment.service'
 @ApiBasicAuth()
 export class CommentController implements CommentREST {
     constructor(
-        private readonly commentService: CommentService
+        private readonly commentService: CommentService,
+        @Inject(forwardRef(() => IssueService))
+        private readonly issueService: IssueService,
+        private readonly memberService: MemberService,
+        @Inject(REQUEST)
+        private readonly request: Express.Request
     ) {}
 
     @Get()
@@ -27,6 +35,16 @@ export class CommentController implements CommentREST {
     async addComment(
         @Body() data: CommentAddData
     ): Promise<Comment> {
+        if (!data) {
+            throw new NotFoundException()
+        }
+        const issue = await this.issueService.getIssue(data.issueId)
+        if (!issue) {
+            throw new NotFoundException()
+        }
+        if ((await this.memberService.findMembers(issue.productId, (<User> this.request.user).id)).length == 0) {
+            throw new ForbiddenException()
+        }
         return this.commentService.addComment(data)
     }
 
@@ -36,6 +54,17 @@ export class CommentController implements CommentREST {
     async getComment(
         @Param('id') id: string
     ): Promise<Comment> {
+        const comment = await this.commentService.getComment(id)
+        if (!comment) {
+            throw new NotFoundException()
+        }
+        const issue = await this.issueService.getIssue(comment.issueId)
+        if (!issue) {
+            throw new NotFoundException()
+        }
+        if ((await this.memberService.findMembers(issue.productId, (<User> this.request.user).id)).length == 0) {
+            throw new ForbiddenException()
+        }
         return this.commentService.getComment(id)
     }
 
@@ -46,6 +75,17 @@ export class CommentController implements CommentREST {
     async updateComment(
         @Param('id') id: string, @Body() data: CommentUpdateData
     ): Promise<Comment> {
+        const comment = await this.commentService.getComment(id)
+        if (!comment) {
+            throw new NotFoundException()
+        }
+        const issue = await this.issueService.getIssue(comment.issueId)
+        if (!issue) {
+            throw new NotFoundException()
+        }
+        if ((await this.memberService.findMembers(issue.productId, (<User> this.request.user).id)).length == 0) {
+            throw new ForbiddenException()
+        }
         return this.commentService.updateComment(id, data)
     }
 
@@ -55,6 +95,17 @@ export class CommentController implements CommentREST {
     async deleteComment(
         @Param('id') id: string 
     ): Promise<Comment> {
+        const comment = await this.commentService.getComment(id)
+        if (!comment) {
+            throw new NotFoundException()
+        }
+        const issue = await this.issueService.getIssue(comment.issueId)
+        if (!issue) {
+            throw new NotFoundException()
+        }
+        if ((await this.memberService.findMembers(issue.productId, (<User> this.request.user).id)).length == 0) {
+            throw new ForbiddenException()
+        }
         return this.commentService.deleteComment(id)
     }
 }
