+++ packages/backend/src/rest/versions/version.controller.ts
@@ -1,19 +1,23 @@
 import 'multer'
-import { Body, Controller, Delete, Get, Param, Post, Put, Query, UploadedFile, UseGuards, UseInterceptors } from '@nestjs/common'
+import { Body, Controller, Delete, ForbiddenException, Get, Inject, NotFoundException, Param, Post, Put, Query, UploadedFile, UseGuards, UseInterceptors } from '@nestjs/common'
 import { FileInterceptor } from '@nestjs/platform-express'
 import { ApiBody, ApiResponse, ApiParam, ApiQuery, ApiBasicAuth } from '@nestjs/swagger'
-import { Version, VersionAddData, VersionREST } from 'productboard-common'
+import { User, Version, VersionAddData, VersionREST } from 'productboard-common'
 import { VersionService } from './version.service'
 import { AuthGuard } from '@nestjs/passport'
+import { MemberService } from '../members/member.service'
+import { REQUEST } from '@nestjs/core'
 
 @Controller('rest/versions')
 @UseGuards(AuthGuard('basic'))
 @ApiBasicAuth()
 export class VersionController implements VersionREST<string, Express.Multer.File> {
     constructor(
-        private versionService: VersionService
+        private versionService: VersionService,
+        private readonly memberService: MemberService,
+        @Inject(REQUEST)
+        private readonly request: Express.Request
     ) {}
-
 
     @Get()
     @ApiQuery({ name: 'product', type: 'string', required: true })
@@ -21,6 +25,9 @@ export class VersionController implements VersionREST<string, Express.Multer.Fil
     async findVersions(
         @Query('product') product: string
     ): Promise<Version[]> {
+        if ((await this.memberService.findMembers(product, (<User> this.request.user).id)).length == 0) {
+            throw new ForbiddenException()
+        }
         return this.versionService.findVersions(product)
     }
 
@@ -32,7 +39,11 @@ export class VersionController implements VersionREST<string, Express.Multer.Fil
         @Body('data') data: string,
         @UploadedFile() file: Express.Multer.File
     ): Promise<Version> {
-        return this.versionService.addVersion(JSON.parse(data), file)
+        const dataParsed = <VersionAddData> JSON.parse(data)
+        if ((await this.memberService.findMembers(dataParsed.productId, (<User> this.request.user).id)).length == 0) {
+            throw new ForbiddenException()
+        }
+        return this.versionService.addVersion(dataParsed, file)
     }
 
     @Get(':id')
@@ -41,7 +52,14 @@ export class VersionController implements VersionREST<string, Express.Multer.Fil
     async getVersion(
         @Param('id') id: string
     ): Promise<Version> {
-        return this.versionService.getVersion(id)
+        const version = await this.versionService.getVersion(id)
+        if (!version) {
+            throw new NotFoundException()
+        }
+        if ((await this.memberService.findMembers(version.productId, (<User> this.request.user).id)).length == 0) {
+            throw new ForbiddenException()
+        }
+        return version
     }
 
     @Put(':id')
@@ -54,7 +72,13 @@ export class VersionController implements VersionREST<string, Express.Multer.Fil
         @Body('data') data: string,
         @UploadedFile() file?: Express.Multer.File
     ): Promise<Version> {
-        console.log(typeof data)
+        const version = await this.versionService.getVersion(id)
+        if (!version) {
+            throw new NotFoundException()
+        }
+        if ((await this.memberService.findMembers(version.productId, (<User> this.request.user).id)).length == 0) {
+            throw new ForbiddenException()
+        }
         return this.versionService.updateVersion(id, JSON.parse(data), file)
     }
 
@@ -64,6 +88,13 @@ export class VersionController implements VersionREST<string, Express.Multer.Fil
     async deleteVersion(
         @Param('id') id: string
     ): Promise<Version> {
+        const version = await this.versionService.getVersion(id)
+        if (!version) {
+            throw new NotFoundException()
+        }
+        if ((await this.memberService.findMembers(version.productId, (<User> this.request.user).id)).length == 0) {
+            throw new ForbiddenException()
+        }
         return this.versionService.deleteVersion(id)
     }
 }
